push SLOT checks into the main qlist_match loop #353223 by Pacho Ramos
the current logic does a match on CAT/PN-PVR, and then if that works, does
a compare on the slots of the two values.  this works if the input arg is
unique w/out the SLOT, but fails otherwise.

for example, if you have installed:
	sys-devel/gcc-4.7.0 (SLOT=4.7.0)
	sys-devel/gcc-4.7.1 (SLOT=4.7.1)
	sys-devel/gcc-4.7.2 (SLOT=4.7.2)

if the user runs:
	qlist -IvS gcc-4.7.{0,1,2}
they'll get correct results (because the user args sans SLOTs are unique)

but if they run:
	qlist -IvS gcc:4.7.{0,1,2}
they'll only get one result (gcc:4.7.0) because the user args sans SLOTs
are all the same ("gcc").  so the code will match the installed gcc-4.7.0
against all the args, but then fail the SLOT check for the latter two.

in order to fix this, we have to merge the CAT/PN-PVR checks with the SLOT
checks so they're done at the same time.  i.e. qlist_match() cannot match
an atom if the input includes a SLOT qualifier and rely on external checking.