vdb: add a q_vdb_pkg_eat helper to unify various call styles
Now that the core eat func is based on dynamic bufs, it's a lot easier
to create a core vdb helper for eating pkg files.

This has the added advantage of making the pkg dir open a delayed call.
We only opendir when we actually want to read files out of it.  For many
funcs which don't do that, it saves a lot of syscall overhead.