which: punt!
This code is only used by --install, and only when /proc/self/exe does
not work.  We rarely utilize --install, and it's rare for /proc to be
broken in a way we can't rely on.  So having this func just for that
does not make much sense.

Even then, the code was not correct.  It walked $PATH in reverse order
(when it should have been forward order), and it would abort scanning
beofre it checked the first element.  It also doesn't support empty
path elements (which is supposed to be $PWD).

If we want a which() in the future, we can grab the updated version
from Gentoo's pax-utils project.