cache: rework and move to /var/cache/edb
Since we want the portdir to be read-only when possible, move our
cache files to /var/cache/edb.  This means the cache regeneration
logic isn't as robust (as we can't rely on rsync deleting it), but
so be it.  We can see about improving this in the future.

URL: https://bugs.gentoo.org/548052
Reported-by: Martin Väth <martin@mvath.de>