xsystem: fix fchdir handling with <linux-3.5
We use O_PATH with the vdb dirs, and sometimes we try to fchdir with
those fds.  Unfortunately, fchdir doesn't work with O_PATH and linux-3.5
leading to errors like:
	>>> pkg_postrm
	merge: fchdir failed: Bad file descriptor

When fchdir fails with EBADF, assume it's because of an older kernel,
and fall back to using the /proc/self/fd/ path indirectly.