support repos.conf files
Since the system is moving away from PORTDIR/etc... and to a more
flexible repos.conf setup, make sure we support that stuff.

This pulls in an external dependency which is new for us, but the
iniparser library is very small (<20k on x86_64), written in pure
C, and is used by big projects already (like samba).  Doing an NIH
thing here would be a waste of time.

URL: https://bugs.gentoo.org/540620
Reported-by: Elias Probst <mail@eliasprobst.eu>