qatom: add a --format flag for controlling output
Rather than forcing people to deal with the ad-hoc output and run it
through another program to reassemble things, add a --format flag so
they can explicitly get the output hey care about.

This drops support for the letter field, but no one was really using
that in the first place.  If anyone does care, we can re-add it as a
new field.

This does change the output slightly from existing behavior: there
will be extra spaces between the latter fields, but this shouldn't
break existing programs as they generally split on whitespace.