qdepends: rework dep tree flattening
By leveraging stpcpy, we don't need to call strlen at all, or update the
length of the buffer we've written to.  Instead, make the func recursive
and pass around the start of the buffer that is ready to accept more data.

While we're here, increase the static buffer to 1MiB.  Some packages (e.g.
qemu) are larger than the 8KiB we have which causes corruption/crashes.
There's no way we'd exceed 1MiB! </last-words>