quse: fix invalid close
This happens to work normally as we only have fds {0,1,2} open
and {3} is the active file we're working on.  Since the f loop
contains 3 (for an unrelated reason), we close the right fd.
But if there are more/fewer fd's open at start, it might fail,
and we leak the FILE* structure.  Switch to closing the right
file pointer to avoid all this mess.