fix minor overlay memory leak w/repos.conf parsing
When repos.conf is active, we end up leaking the default main overlay
memory.  Make sure we free it if repos.conf was loaded.

Also drop the old closing of stderr so asan can write its output.  The
addition of this call was due to a bug in the old stderr reopen logic
that was rewritten in commit ba5f0328.