qatom: print unset members as <unset> instead of relying on libc
Some libcs do NOT do the favour of turning a NULL-pointer into the
string "(null)", but segfault instead.  This is just a friendly guesture
of libcs which do, but it is nowhere standard or that this should result
in "(null)".  While "(null)" is actually a developer concept, and
possibly out of context for users, check for printing NULL-pointers and
replace them with "<unset>" instead.