libq/tree: introduce a wrapper for retrieving ebuild metadata
To further hide the difference between trees, add a wrapper around
tree_get_meta, which deals with the underlying storage format.  This is
most notably useful when dealing with vdb-based and binpkgs because they
have a specific or more expensive way of retrieval.  This can now
properly be hidden from the q applets.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>