qsize: report the actual disk usage of packages
Files installed by some packages are using hard links in order
to save disk space. In such a case, the real size of the package
on the disk isn't a plain sum of the sizes of its files, because
multiple paths map to the same inode number.

This patch modifies qsize to report the actual disk usage of a package
by summing the sizes of unique inodes belonging to the package.

The implementation uses libq/set, which performs reasonably
well even for packages containing thousands of files.
Tested on an ext4 filesystem.

Before:
    $ qsize mesa
    media-libs/mesa: 136 files, 31 non-files, 2.2G

After:
    $ qsize mesa
    media-libs/mesa: 136 files, 31 non-files, 1012.1M