qmerge: use libq/tree for binpkg/Packages needs
Instead of manual parsing of Packages file and traversing binpkgs, use
the functionality provided by libq/tree, most notably tree_match_atom().

This is just phase one, where we replace the custom code with libq/tree
code and then fill in the Pkg structure.  In a next phase we should skip
this and directly use tree_match_ctx, with atom and meta in there.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>