tests: add qmerge tests for different compression schemes in binpkgs
While we had support for installing various compressed binpkgs, we never
tested this at all.  Especially since zstd now became the new standard
Portage uses, it is a good idea to ensure we cope well with this.

This new blob of tests, checks qtbz2, qlist and qmerge to disassemble,
assemble and install binpkgs with all known supported compressors.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>