qlist: fix matching of USE-flags
Thanks to Michael Yagliyan for finding this bug and proposing the fix.

We cannot perform merge-sort if the comparison in use is different
between sorting and processing afterwards.  Use case-sensitive sort
everywhere, for it is cheaper, matches Portage and is safer/more
correct/inline with PMS.  This does change the default output ordering
of the flags though.

Added a test that ensures capital USE-flag is now matched correctly.

Bug: https://bugs.gentoo.org/829579
Signed-off-by: Fabian Groffen <grobian@gentoo.org>