man: fix -S/--slots option description in qlist.1
I noticed that the qlist(1) man page did not mention that -S implies -I,
so I wanted to fix that.

The man page was actually supposed to mention it, but
man/include/qlist.optdesc.yaml was incorrectly using the "slot" key to
define --slots's long description, so the man page ended up using the
short description from  q list --help  instead of the longer description
that contains "This flag implies -I.".

Renaming the key in the yaml file from "slot" to "slots" fixed the
problem.

I have regenerated the man pages with man/mkman.py, but only included
the new version of qlist.1 in this commit.

Signed-off-by: Emanuele Torre <torreemanuele6@gmail.com>