main.c: fix ROOT handling after b7acc09ff2d9d2ce176ee131458ad39489b0bbd3
Previous fixes caused errors like:

qwhich: tree_open(): could not open repository at var/db/repos/gentoo
(under root /)

The proposed fix as in PR #31 was not correct because portroot is
already resolved.  Instead some assumptions here and there incorrectly
relied on paths being absolute or not.

This is and remains messy, in particular when the combined path (in
ROOT) doesn't exist, which means realpath will fail, and we need to go
with whatever input we had.

Closes: https://github.com/gentoo/portage-utils/pull/31
Signed-off-by: Fabian Groffen <grobian@gentoo.org>