libq/tree: consider ebuild-tree cache as a cache, not a tree
While it is certainly faster to only consider a metadata cache (such as
md5-cache) it yields incorrect results on trees in flux or under
development.  The only correct way is to consider the ebuilds, and look
at an existing cache, if it matches.

This commit implements this behaviour for libq/tree, used by various q
utilities.  It considers an cache for an ebuild tree, and uses it for
operations where it needs info from inside the ebuilds (such as DEPEND,
etc.) over trying to fetch it from the ebuild itself.

This is considerably slower than just considering the metadata alone,
if only because we now have to traverse multiple extra directories (one
for each package) as opposed to just reading the metadata.  However,
this is unavoidable, and hopefully can be mitigated by parallel scans in
the future.

Bug: https://bugs.gentoo.org/934549
Bug: https://bugs.gentoo.org/898194
Signed-off-by: Fabian Groffen <grobian@gentoo.org>