q: introduce first iteration of gtree "cache"
Not really a cache, but a single container (tar), compressed (zstd)
containing all ebuilds, their metadata and eclasses.  The format is
somewhat inspired by GLEP-78 GPKG support, but without the nested
archives.  If the signature should be contained, it probably should
become like that, which is why the format is not yet "finished".
A full tree in gtree-1 format is 8.1MiB (uncompressed 250MiB) and should
allow dep-resolving or metadata lookups more efficiently than opening up
many files.  The starting point is that everything is consistant, so no
more checks are required to ensure cache and data are in sync.  Another interesting feature of the gtree-1 container is that it could be easily
retrieved over https and used to compute and install updates.  The
format allows to use the same layout for VDB or binpkg/packages, but
this angle is not fully explored as of now.

Next commits should bring support for gtree-1 in libq/tree to
transparantly utilise the "cache" when available.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>