q: drop entries for cache in gtree
It appears archive entries are expensive in storage, so cut back by
stashing all of the cache entries in a single file, much like md5-cache
does.  There appears to be no use for these separately that would be
seriously punished by having to grab all of them together (and parse the
rows).

Signed-off-by: Fabian Groffen <grobian@gentoo.org>