q: re-order files in gtree cache
Separate cache from ebuilds, and when dealing with ebuilds, don't repeat
metadata, Manifest, files/ for every ebuild.  The latter only saves a
little (approx. 1MiB) but it is cleaner from the extraction point of
view.  We can now sort of restore this part of the original tree.

Store cache entries first, per PF, such that a process only interested
in the cache (which is most of the time the case) can stop reading
after.  Similarly, the sorting of entries allows a reader to determine
early a requested atom is not available in the gtree archive.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>