*: drop usage of dirent d_type
d_type is not defined by POSIX in dirent, but in posix_getdent.  The
latter is not available on most platforms, and gnulib doesn't provide a
wrapper.  Realistically, using stat is portable, and the places in which
we really need to know the type are very limited.  So drop using d_type.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>