q: implement rudimentary jobserver
Allow q to run as jobserver for systems/scenarios where better solutions
like steve are not possible.  This mostly targets Gentoo Prefix systems,
which by use of this simple jobserver can still achieve a global
parallelism threshold for parallel make invocations, especially when
Portage's support for this allows to get the token administration
correct (as to help with Make's assumed implicit token per startup).

The jobserver is canonically available in the filesystem via:
  EPREFIX/tmp/.q-jobserver.sock
and as such can be statically used in e.g. MAKEFLAGS with
  --jobserver-auth=fifo:EPREFIX/tmp/.q-jobserver.sock

Signed-off-by: Fabian Groffen <grobian@gentoo.org>