qmerge: fix handling gpkg binpkgs with top-level dirs defined
Tar archives don't need directories to be defined, but they can define
them, for the top-level ones, which we ignore when unpacking metadata
and image, ignore it if they are present.

While at it, improve several error messages to be unique so they will
point to the right location in the code.

Bug: https://bugs.gentoo.org/968185#c5
Signed-off-by: Fabian Groffen <grobian@gentoo.org>